feat: Implement CacheService for improved performance across modules
