+++ modules/AnimeStatsService.js
@@ -2,45 +2,11 @@ const axios = require('axios');
 const { EmbedBuilder } = require('discord.js');
 const logger = require('../logger');
 const metricsService = require('../metrics');
-
-// Caching Service
-class CacheService {
-    constructor(ttl = 300000) { // 5 minutes default TTL
-        this.cache = new Map();
-        this.ttl = ttl;
-    }
-
-    set(key, value) {
-        const entry = {
-            value,
-            timestamp: Date.now()
-        };
-        this.cache.set(key, entry);
-        return value;
-    }
-
-    get(key) {
-        const entry = this.cache.get(key);
-        if (!entry) return null;
-
-        // Check if entry is expired
-        if (Date.now() - entry.timestamp > this.ttl) {
-            this.cache.delete(key);
-            return null;
-        }
-
-        return entry.value;
-    }
-
-    clear(key) {
-        this.cache.delete(key);
-    }
-}
+const CacheService = require('./CacheService');
 
 // Main Logic
 class AnimeStatsService {
     constructor() {
-        // Remove the accessTokenFn parameter
         this.cache = new CacheService();
     }
